requirements.txt
numpy
einops
thop
pdb
scikit-image
pillow
tqdm